posts-service/requirements.txt
requests
datetime
flask
psycopg2
flask_cors
Flask-SocketIO==4.3.1
python-engineio==3.13.2
python-socketio==4.6.0
eventlet
redis
uwsgi
gevent
gunicorn
celery
redis
pyJWT